+++ FluentTerminal.App.Services/Terminal.cs
@@ -65,7 +65,7 @@ namespace FluentTerminal.App.Services
 
         public string FallbackTitle { get; private set; }
 
-        public int Id { get; }
+        public byte Id { get; }
 
         /// <summary>
         /// To be called by either view or viewmodel
