+++ FluentTerminal.Models/TerminalOutput.cs
@@ -2,7 +2,7 @@
 {
     public class TerminalOutput
     {
-        public int TerminalId { get; set; }
+        public byte TerminalId { get; set; }
         public byte[] Data { get; set; }
     }
 }
